Lista 3 de exercícios em Python

+++ Lista3-Python/ex06-contagio.py
@@ -15,7 +15,7 @@ while (contagemPessoas/populacao*100) < imunidade:
         print(cont)
         break
     if pessoasInfectadas != 0:
-        pessoasInfectadas = int((pessoasInfectadas*txContagio))
+        pessoasInfectadas = (pessoasInfectadas*txContagio)
         contagemPessoas += pessoasInfectadas
     else:
         pessoasInfectadas = qntDia0
